Separate out init step
